% Generated by roxygen2 (4.0.1): do not edit by hand
\name{caret_sis}
\alias{caret_sis}
\title{Find Optimal Subset of Predictors to pass along to the Modeling Stage.  Based on CaretPackage RFE}
\usage{
caret_sis(dataframe, exclude = character(0), parallelize = TRUE)
}
\arguments{
\item{dataframe}{from which to run feature selection upon}

\item{exclude}{character. A vector of variables to never remove.}
}
\description{
Find Optimal Subset of Predictors to pass along to the Modeling Stage.  Based on CaretPackage RFE
}

