% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-datetime.r
\name{datetime_fn}
\alias{datetime_fn}
\title{A simpler, faster mungebit that takes Postgresql CreatedAt statements in the format of YYYY-MM-DD HH:MM:SS.SSSS and convert them into the choice of...
   days since 1970 (default)
   the hour of day (0-23, use mode="hod")
   the day of week (Mon-Sun, use mode="dow")
   the day of month (1-31, use mode="dom")
   the day of year (1-365, use mode="doy")
   the month of year (1-12 corresponding to Jan-Dec, use mode="moy") 
   if the output is a holiday (\code{TRUE} or \code{FALSE}, use mode="holiday")
   if the output is a weekend (\code{TRUE} or \code{FALSE}, use mode="weekend")
   if the output is a business day (\code{TRUE} or \code{FALSE}, use mode="bizday")}
\usage{
datetime_fn(createdat, mode = "since", permissive = FALSE)
}
\arguments{
\item{mode}{gives the desired output format (see above)}

\item{permissive}{logical. If false, will error with an improper type.  If true, will just warn.}

\item{date}{contains the date to be formatted}
}
\description{
A simpler, faster mungebit that takes Postgresql CreatedAt statements in the format of YYYY-MM-DD HH:MM:SS.SSSS and convert them into the choice of...
   days since 1970 (default)
   the hour of day (0-23, use mode="hod")
   the day of week (Mon-Sun, use mode="dow")
   the day of month (1-31, use mode="dom")
   the day of year (1-365, use mode="doy")
   the month of year (1-12 corresponding to Jan-Dec, use mode="moy") 
   if the output is a holiday (\code{TRUE} or \code{FALSE}, use mode="holiday")
   if the output is a weekend (\code{TRUE} or \code{FALSE}, use mode="weekend")
   if the output is a business day (\code{TRUE} or \code{FALSE}, use mode="bizday")
}

