\name{discretizer}
\alias{discretizer}
\alias{discretizer_fn}
\title{Discretizer}
\usage{
discretizer(dataframe, input_cols = colnames(dataframe), ...)
}
\arguments{
  \item{dataframe}{a data.frame to discretize.}

  \item{cols}{a vector of columns to discretize.}

  \item{...}{the arguments passed to the discretization.}
}
\description{
Discretizer
}

