% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drop_single_value_variables.r
\name{drop_single_value_variables_fn}
\alias{drop_single_value_variables_fn}
\title{Drop variables with only one unique value}
\usage{
drop_single_value_variables_fn(x)
}
\arguments{
\item{x}{an atomic vector.}
}
\description{
Drop variables with only one unique value
}

