% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drop_variables_fn}
\alias{drop_variables_fn}
\title{Drop variables in a dataframe.}
\usage{
drop_variables_fn(x)
}
\arguments{
\item{x}{an atomic vector. The column to process.}
}
\value{
NULL ensures the column gets removed
}
\description{
This is a column transformation and will be wrapped in
\code{column_transformation}.
}
\examples{
drop_variables(iris, 1) # Drop first column
drop_variables(iris, c('Sepal.Length', 'Petal.Length'))
drop_variables(iris, c(TRUE,TRUE,FALSE,FALSE,TRUE)) # only leave columns 3 and 4
}

