% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imputer2}
\alias{imputer2}
\title{Imputer2}
\usage{
imputer2(dataframe)
}
\arguments{
\item{dep_var_name}{Name of the dependent variable.  (Default is "dep_var".)}

\item{maxlevels}{Maximum number of levels allowed for a factor variable to be imputed.  Too many levels would cause the mungebit to be very slow.}
}
\description{
Impute columns by fitting models to other columns in data set
Numeric variables are imputed by a LASSO regression, while
factor variables are imputed using LDA.  The imputation works
only in training mode because it is difficult to guarantee that
an arbitrary test set will be correctly predicted by the lars
and lda functions.  (For example, if there are new levels
of a factor variable in the test set, the model matrix during
prediction will have a different number of columns than during
training, causing the lars algorithm to fail.)
}

