% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.r
\name{imputer_fn}
\alias{imputer_fn}
\title{Impute a column by mean
Only meant to be used as a mungebit}
\usage{
imputer_fn(x)
}
\arguments{
\item{x}{an atomic vector.}
}
\description{
Impute a column by mean
Only meant to be used as a mungebit
}

