% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/syberiaMungebits-package.r
\docType{data}
\name{iris_discretized}
\alias{iris_discretized}
\title{A discretized version of the iris data set, used for internal
unit testing.}
\format{A data frame with 150 rows and 5 variables, all factors.}
\usage{
data(iris_discretized)
}
\description{
A discretized version of the iris data set, used for internal
unit testing.
}
\keyword{internal}

