% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.weekend}
\alias{is.weekend}
\title{Take input in a wide variety of date/time formats and output it in a variety of consistent formats.}
\usage{
is.weekend(date)
}
\arguments{
\item{input}{contains the date to be formatted

Makes five assumptions about dates:
 - 12/11/1991 will be December 11, 1991 (US dates)
 - two-digit years > 20 refer to 20th century ('91 refers to 1991)
 - two-digit years <= 20 refer to 21st century ('14 refers to 2014)
 - two-digit years will be at end (12/11/20 will refer to December 11, 1920)
 - numeric strings are dates if length 8, days since 1970 otherwise
 - numeric strings as dates will be YYYYMMDD}

\item{mode}{tells the desired output format
"date" will return a R date
"numeric" will return the number of days since Jan 1, 1970
"holiday" will return TRUE if it's a US federal holiday and FALSE if not
"weekend" will return TRUE if it's a weekend and FALSE if not
"businessday" will return TRUE if it's a weekend or holiday}
}
\description{
Take input in a wide variety of date/time formats and output it in a variety of consistent formats.
}

