% Generated by roxygen2 (4.0.0): do not edit by hand
\name{missing_indicator}
\alias{missing_indicator}
\title{Create a separate indicator column if the numeric variable has a missing level}
\usage{
missing_indicator(dataframe, input_cols = colnames(dataframe),
  output_cols = input_cols, ...)
}
\description{
Create a separate indicator column if the numeric variable has a missing level
}

