% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/new_variable.R
\name{new_variable}
\alias{new_variable}
\title{Creates a new variable in the dataframe.}
\usage{
new_variable(df, fun, output, inputs = NULL)
}
\arguments{
\item{df}{dataframe. The dataframe to modify.}

\item{fun}{function. The function to run to convert inputs to outputs.}

\item{output}{character. The name of the desired new variable.}

\item{inputs}{character. The names of the variables needed to make the output.}
}
\description{
Creates a new variable in the dataframe.
}

