% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/orderer.r
\name{orderer}
\alias{orderer}
\title{Order a dataframe by a column}
\usage{
orderer(dataframe, column_name)
}
\arguments{
\item{dataframe}{a data.frame to re-order}

\item{column_name}{the column by which to order}
}
\description{
Order a dataframe by a column
}

