% Generated by roxygen2 (4.0.2): do not edit by hand
\name{paramatch}
\alias{paramatch}
\title{Mungebit to count the occurances of the N most repeated words in a dataframe column.}
\usage{
paramatch(dataframe, col, top_n_words = 5, suppress.input = FALSE)
}
\arguments{
\item{top_n_words}{integer.  The number of top strings to count in each line.}

\item{suppress.input}{logical.  Suppresses the Input column of the output data frame.}

\item{input}{an atomic vector. The paragraph of text to analyse.}
}
\value{
numeric. Number of occurances of the particular word in each line.
}
\description{
Trains to find the most repeated words.  Then counts the number of matches for dataframes passed to it.
}
\details{
This conducts full word matching only.  Substring matches do not count towards the total.
}
\examples{
paragraph <- data.frame(text = c("This is an example.", "Great code uses examples like this example."), stringsAsFactors = FALSE)
paramatch(paragraph, "text", 2)
}
\author{
Mike Bruno
}

