% Generated by roxygen2 (4.0.0): do not edit by hand
\name{remember}
\alias{remember}
\title{remembers information as of a given step in the mungeprocedure}
\usage{
remember(dataframe, f, key)
}
\arguments{
\item{f}{function of the form f(dataframe) that returns the statistic of interest, e.g. f <- function(x) nrows(x)}

\item{key}{binding for the statistic of interest}
}
\description{
remembers information as of a given step in the mungeprocedure
}
\details{
the result of apply f to dataframe is stored in a list called "memory" in the base environment
}
\examples{
remember(df,dim,"dimensions")
> memory
$dimensions
[1] 454 463
}

