% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove_monotonic.r
\name{remove_monotonic}
\alias{remove_monotonic}
\title{Drop one of each monotonically related pair}
\usage{
remove_monotonic(dataframe, threshold, verbose = FALSE)
}
\arguments{
\item{dataframe}{A data.frame.  Numeric columns must already be set to numeric with
the appropriate mungebit.  (Non-numeric columns will not be considered).  Also only
only complete cases will be considered, so imputation should already have occurred.}

\item{threshold}{If the Spearman rank correlation exceeds this value, columns will be
dropped.}
}
\description{
Drop one of each monotonically related pair
}

