% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/replace_variable.R
\name{replace_variable}
\alias{replace_variable}
\title{Replaces a variable in the dataframe.}
\usage{
replace_variable(df, fun)
}
\arguments{
\item{df}{dataframe. The dataframe to modify.}

\item{fun}{function. The function to run to convert the input into the new input.}
}
\description{
Replaces a variable in the dataframe.
}

