% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sampler}
\alias{sampler}
\title{Take a data frame and upsample a single segment
i.e. Given a variable var and a value val, randomly sample with replacement the rows with var == val}
\usage{
sampler(df, var, val, frac)
}
\arguments{
\item{df}{a data frame}
}
\description{
Take a data frame and upsample a single segment
i.e. Given a variable var and a value val, randomly sample with replacement the rows with var == val
}

