\name{sure_independence_screen}
\alias{sure_independence_screen}
\title{Sure independence screening function}
\usage{
sure_independence_screen(dataframe, ...)
}
\arguments{
  \item{dataframe}{a data.frame. Run sure independence
  screening on this dataframe.}

  \item{...}{further SIS options. See also
  \code{sure_independence_screening} in the
  \code{statsUtils} package.}
}
\description{
Sure independence screening function
}

