% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/truncator.r
\name{truncator_fn}
\alias{truncator_fn}
\title{Constrain a numeric column to the range [min,max] of the training set}
\usage{
truncator_fn(x)
}
\arguments{
\item{x}{an atomic vector}

\item{digits}{how many decimal places you would like your numeric number to contain after its been truncated}
}
\description{
Constrain a numeric column to the range [min,max] of the training set
}

