% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/visualize.r
\name{visualize}
\alias{visualize}
\title{Attain Visual insights into data. Plot overlayed histogram vs dep_var for each variable in dataframe}
\usage{
visualize(dataframe, ...)
}
\arguments{
\item{dataframe}{a data.frame}

\item{...}{excess arguments one can pass to the overlayed_histogram function. I.E. line_var, histColor, etc.}
}
\description{
Attain Visual insights into data. Plot overlayed histogram vs dep_var for each variable in dataframe
}

