% Generated by roxygen2 (4.0.2): do not edit by hand
\name{date_parser_fn}
\alias{date_parser_fn}
\title{Parse numeric date distances into numerics.}
\usage{
date_parser_fn(column)
}
\arguments{
\item{column}{an atomic vector. The column to process.}
}
\value{
numeric. Number of years
}
\description{
For example, c("8 days", "100 days", "") will become c(8, 100, NA)
}
\details{
This is a column transformation and will be wrapped in
\code{column_transformation}.
}
\examples{
dd <- data.frame(cbind(c("", "8 days", "20 days")), stringsAsFactors = FALSE)
date_parser(dd, 1)
print(dd[[1]])
# [1] 0 8 20
}

