% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discretizer.r
\name{discretizer}
\alias{discretizer}
\title{Discretizer}
\usage{
discretizer(dataframe, input_cols = colnames(dataframe), ...)
}
\arguments{
\item{dataframe}{a data.frame to discretize.}

\item{input_cols}{a vector of columns to discretize.}

\item{...}{the arguments passed to the discretization.}
}
\description{
Discretizer
}

