% Generated by roxygen2 (4.0.0): do not edit by hand
\name{discretizer}
\alias{discretizer}
\title{Discretizer}
\usage{
discretizer(dataframe, input_cols = colnames(dataframe), ...)
}
\arguments{
\item{dataframe}{a data.frame to discretize.}

\item{input_cols}{a vector of columns to discretize.}

\item{...}{the arguments passed to the discretization.}
}
\description{
Discretizer
}

