% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drop_percent_missing.r
\name{drop_percent_missing}
\alias{drop_percent_missing}
\title{Drop columns with over a certain percent missing.}
\usage{
drop_percent_missing(dataframe, input_cols = colnames(dataframe), ...)
}
\arguments{
\item{dataframe}{a data.frame to drop columns from.}

\item{input_cols}{a vector of columns to drop percent missing from.}

\item{...}{the arguments passed to the transformation (e.g., \code{threshold}).}
}
\description{
Drop columns with over a certain percent missing.
}
\seealso{
\code{\link{drop_percent_missing_fn}}
}

