% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imputer_fn}
\alias{imputer_fn}
\title{Impute a column by mean
Only meant to be used as a mungebit}
\usage{
imputer_fn(x)
}
\arguments{
\item{x}{an atomic vector.}
}
\description{
Impute a column by mean
Only meant to be used as a mungebit
}

