% Generated by roxygen2 (4.0.2): do not edit by hand
\name{paramatch}
\alias{paramatch}
\title{Count the occurances of the N most repeated words in a paragraph}
\usage{
paramatch(input, top_n_words = 5, suppress.input = FALSE)
}
\arguments{
\item{input}{an atomic vector. The paragraph of text to analyse.}

\item{top_n_words}{integer.  The number of top strings to count in each line.}

\item{suppress.input}{logical.  Suppresses the Input column of the output data frame.}
}
\value{
numeric. Number of occurances of the particular word in each line.
}
\description{
Paragraphs of M lines for N words will return a data frame of N + 1 columns and M rows each counting the occurance in the particular line.
}
\details{
This conducts full word matching only.  Substring matches do not count towards the total.
}
\examples{
paragraph <- c("This is an example.", "Great code uses examples like this example.")
paramatch(paragraph, 2)
}

