% Generated by roxygen2 (4.0.2): do not edit by hand
\name{paramatch}
\alias{paramatch}
\title{Mungebit to count the occurances of the N most repeated words in a dataframe column.}
\usage{
paramatch(dataframe, col, top_n_words = 5, suppress.input = FALSE,
  blacklist = c())
}
\arguments{
\item{dataframe}{}

\item{col}{character. The column name pertaining to paragraphs for matching.}

\item{top_n_words}{integer.  The number of top strings to count in each line.}

\item{suppress.input}{logical.  Suppresses the Input column of the output data frame.}

\item{blacklist}{character. Prevents undesired words participating int he output.}
}
\value{
numeric. Number of occurances of the particular word in each line.
}
\description{
Trains to find the most repeated words.  Then counts the number of matches for dataframes passed to it.
}
\details{
This conducts full word matching only.  Substring matches do not count towards the total.
}
\examples{
mp <- mungebits::mungeplane(data.frame(id = c(1:2), text = c("This is an example.", "Great code uses examples like this example."), stringsAsFactors = FALSE))
mb <- mungebits:::mungebit(paramatch)
mb$run(mp, col = 'text', top_n = 2)
}
\author{
Mike Bruno
}

