% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/paramatch.r
\name{paramatch}
\alias{paramatch}
\title{Identify and Count occurances of the most repeated words.}
\usage{
paramatch(dataframe, col, top_n = 5, suppress.input = FALSE,
  blacklist = c())
}
\arguments{
\item{dataframe}{}

\item{col}{character. The column name pertaining to paragraphs for matching.}

\item{top_n}{integer.  The number of top strings to count in each line.}

\item{suppress.input}{logical.  Suppresses the Input column of the output data frame.}

\item{blacklist}{character. Prevents undesired words participating int he output.}
}
\value{
numeric. Number of occurances of the particular word in each line.
}
\description{
Finds the N the most repeated words in a character column of a dataframe and appends N columns containing the number of matches of each word in each row.
}
\details{
This conducts full word matching only.  Substring matches do not count towards the total.
}
\examples{
mp <- mungebits::mungeplane(data.frame(id = c(1:2), text = c("This is an example.", "Great code uses examples like this example."), stringsAsFactors = FALSE))
mb <- mungebits:::mungebit(paramatch)
mb$run(mp, col = 'text', top_n = 2)
}
\author{
Mike Bruno
}

