% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_outliers_fn}
\alias{remove_outliers_fn}
\title{remove outliers based upon a z_score threshold z_score=> (x - mean(x) / sd(x)}
\usage{
remove_outliers_fn(x, threshold = 3)
}
\arguments{
\item{x}{an numeric vector}

\item{threshold}{a z_score threshold from which the absolute value of a z_score above this threshold will be set to NA}
}
\description{
remove outliers based upon a z_score threshold z_score=> (x - mean(x) / sd(x)
}

