% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/restore_categorical_variables.R
\name{restore_categorical_variables_fn}
\alias{restore_categorical_variables_fn}
\title{Restores categorical variables, allowing you to predict.}
\usage{
restore_categorical_variables_fn(x)
}
\arguments{
\item{x.}{The column to restore.}
}
\description{
Restores categorical variables, allowing you to predict.
}

